+++ Common/src/main/java/com/tristankechlo/wool_collection/init/ModRegistry.java
@@ -3,9 +3,9 @@ package com.tristankechlo.wool_collection.init;
 import com.tristankechlo.wool_collection.TheWoolCollection;
 import com.tristankechlo.wool_collection.blocks.WeavingStationBlock;
 import com.tristankechlo.wool_collection.container.WeavingStationContainer;
+import com.tristankechlo.wool_collection.platform.IPlatformHelper;
 import com.tristankechlo.wool_collection.platform.RegistrationProvider;
 import com.tristankechlo.wool_collection.platform.RegistryObject;
-import com.tristankechlo.wool_collection.platform.Services;
 import com.tristankechlo.wool_collection.recipe.WeavingStationRecipe;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.inventory.MenuType;
@@ -29,7 +29,7 @@ public final class ModRegistry {
     /* Content */
     public static final RegistryObject<Block> WEAVING_STATION_BLOCK = BLOCKS.register("weaving_station", WeavingStationBlock::new);
     public static final RegistryObject<BlockItem> WEAVING_STATION_ITEM = ITEMS.register("weaving_station", () -> new BlockItem(WEAVING_STATION_BLOCK.get(), new Item.Properties()));
-    public static final RegistryObject<MenuType<WeavingStationContainer>> WEAVING_STATION_CONTAINER = CONTAINERS.register("weaving_station", Services.PLATFORM.buildContainer());
+    public static final RegistryObject<MenuType<WeavingStationContainer>> WEAVING_STATION_CONTAINER = CONTAINERS.register("weaving_station", IPlatformHelper.INSTANCE.buildContainer());
     public static final RegistryObject<RecipeType<WeavingStationRecipe>> WEAVING_STATION_RECIPE_TYPE = RECIPES.register("weaving_station", () -> new RecipeType<>() {});
     public static final RegistryObject<RecipeSerializer<WeavingStationRecipe>> WEAVING_STATION_RECIPE_SERIALIZER = RECIPE_SERIALIZERS.register("weaving_station", WeavingStationRecipe.Serializer::new);