+++ Forge/src/main/java/com/tristankechlo/wool_collection/TheWoolCollectionForge.java
@@ -2,7 +2,8 @@ package com.tristankechlo.wool_collection;
 
 import com.tristankechlo.wool_collection.init.ModBlocks;
 import com.tristankechlo.wool_collection.init.ModRegistry;
-import net.minecraft.world.item.CreativeModeTabs;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.FireBlock;
 import net.minecraft.world.level.block.state.properties.BlockSetType;
@@ -40,21 +41,20 @@ public class TheWoolCollectionForge {
     }
 
     /* add items to their creative tab */
-    private void onCreativeModeTabRegister(CreativeModeTabEvent.BuildContents event) {
-        if (event.getTab() == CreativeModeTabs.COLORED_BLOCKS) {
-            TheWoolCollection.sortedListByColor(ModBlocks.FENCES).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.FENCE_GATES).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.STAIRS).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.SLABS).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.WALLS).forEach(event::accept);
-        }
-        if (event.getTab() == CreativeModeTabs.REDSTONE_BLOCKS) {
-            TheWoolCollection.sortedListByColor(ModBlocks.BUTTONS).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.PRESSURE_PLATES).forEach(event::accept);
-        }
-        if (event.getTab() == CreativeModeTabs.FUNCTIONAL_BLOCKS) {
-            event.accept(ModRegistry.WEAVING_STATION_ITEM.get());
-        }
+    private void onCreativeModeTabRegister(CreativeModeTabEvent.Register event) {
+        event.registerCreativeModeTab(new ResourceLocation(TheWoolCollection.MOD_ID, "general"), builder ->
+                builder.title(Component.translatable("itemGroup.wool_collection.general"))
+                        .icon(() -> ModRegistry.WEAVING_STATION_ITEM.get().getDefaultInstance())
+                        .displayItems((params, output) -> {
+                            output.accept(ModRegistry.WEAVING_STATION_ITEM.get());
+                            TheWoolCollection.sortedListByColor(ModBlocks.FENCES).forEach(output::accept);
+                            TheWoolCollection.sortedListByColor(ModBlocks.FENCE_GATES).forEach(output::accept);
+                            TheWoolCollection.sortedListByColor(ModBlocks.STAIRS).forEach(output::accept);
+                            TheWoolCollection.sortedListByColor(ModBlocks.SLABS).forEach(output::accept);
+                            TheWoolCollection.sortedListByColor(ModBlocks.WALLS).forEach(output::accept);
+                            TheWoolCollection.sortedListByColor(ModBlocks.PRESSURE_PLATES).forEach(output::accept);
+                            TheWoolCollection.sortedListByColor(ModBlocks.BUTTONS).forEach(output::accept);
+                        }));
     }
 
     /* mark all blocks as flammable */